move init config
